// frontend/src/generated/model/pathResultItem.ts
/**
 * backend
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PathResultItem { 
    id: string;
    role: PathResultItem.RoleEnum;
}
export namespace PathResultItem {
    export const RoleEnum = {
        MAIN: 'main',
        OUTER: 'outer'
    } as const;
    export type RoleEnum = typeof RoleEnum[keyof typeof RoleEnum];
}