// frontend/src/generated/model/pairPathStatisticsInterface.ts
/**
 * backend
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PathStatisticsInterface } from './pathStatisticsInterface';


export interface PairPathStatisticsInterface { 
    second: PathStatisticsInterface;
    first: PathStatisticsInterface;
}